# geoimagenet_api/openapi.yaml
openapi: "3.0.0"
info:
  title: Geoimagenet API
  description: API for Geoimagenet
  version: 1.0.0
servers:
  - url: /api/v1
    description: Production server
paths:
...
  /taxonomy_classes:
    get:
      summary: "Taxonomy classes: Search"
      parameters:
        - name: taxonomy_name
          description: >
                       Full name or sluggified name of the taxonomy.
                       Example of name slug for "Couverture de sol": couverture-de-sol
          in: query
          required: true
          schema:
            type: string
        - name: name
          description: Name of the taxonomy class
          required: true
          in: query
          schema:
            type: string
        - name: depth
          in: query
          description: The depth of taxonomy class children to return. -1 means infinite and is the default.
          schema:
            type: integer
            default: -1
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxonomyClass'
        404:
          description: Not found
          content:
            application/json: {}
        400:
          description: Bad request
          content:
            application/json: {}
  /taxonomy_classes/{id}:
    get:
      summary: "Taxonomy classes: Get by id"
      parameters:
        - name: id
          in: path
          description: The id of the taxonomy class
          required: true
          schema:
            type: integer
        - name: depth
          in: query
          description: The depth of taxonomy class children to return. -1 means infinite and is the default.
          schema:
            type: integer
            default: -1
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyClass'
        404:
          description: Not found
          content:
            application/json: {}
  /taxonomy:
    get:
      summary: "Taxonomy: Search"
      parameters:
        - name: name
          in: query
          description: >
                       Full name or sluggified name of the taxonomy.
                       Example of name slug for "Couverture de sol": couverture-de-sol
          schema:
            type: string
        - name: version
          description: The version of the taxonomy. `name` is required if version is provided.
          in: query
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxonomyGroup'
        404:
          description: Not found
          content:
            application/json: {}
        400:
          description: Bad request
          content:
            application/json: {}
        503:
          description: Server is not configured
          content:
            application/json: {}
  /taxonomy/{name_slug}/{version}:
    get:
      summary: "Taxonomy: Get by name slug and version"
      operationId: geoimagenet_api.endpoints.taxonomy.get_by_slug
      parameters:
        - name: name_slug
          in: path
          description: 'Example of name slug for "Couverture de sol": couverture-de-sol'
          required: true
          schema:
            type: string
        - name: version
          in: path
          description: The version of the taxonomy
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Taxonomy'
        404:
          description: Not found
          content:
            application/json: {}
...
components:
  schemas:
...
    TaxonomyClass:
      type: object
      required:
        - id
        - name_fr
        - taxonomy_id
      properties:
        id:
          type: integer
          format: int64
        name_fr:
          type: string
        name_en:
          type: string
        taxonomy_id:
          type: integer
          format: int64
        children:
          type: array
          items:
            $ref: '#/components/schemas/TaxonomyClass'
...
    Taxonomy:
      type: object
      required:
        - id
        - name_fr
        - slug
        - version
      properties:
        id:
          type: integer
          format: int64
        name_fr:
          type: string
        name_en:
          type: string
        slug:
          type: string
        version:
          type: string
        root_taxonomy_class_id:
          type: integer
          format: int64
    TaxonomyGroup:
      type: object
      required:
        - name_fr
        - slug
        - versions
      properties:
        name_fr:
          type: string
        name_en:
          type: string
        slug:
          type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/TaxonomyVersion'
    TaxonomyVersion:
      type: object
      required:
        - taxonomy_id
        - version
      properties:
        taxonomy_id:
          type: integer
          format: int64
        root_taxonomy_class_id:
          type: integer
          format: int64
        version:
          type: string
...